application: add allow_registration config option

+++ app/Http/Controllers/ProfileController.php
@@ -34,8 +34,10 @@ class ProfileController extends Controller
         if ($request->user()->isDirty('email')) {
             $request->user()->email_verified_at = null;
         }
-
-        $request->user()->save();
+        
+        if (config('app.allow_registration')) {
+            $request->user()->save();
+        }
 
         return Redirect::route('profile.edit');
     }
@@ -52,8 +54,10 @@ class ProfileController extends Controller
         $user = $request->user();
 
         Auth::logout();
-
-        $user->delete();
+        
+        if (config('app.allow_registration')) {
+            $user->delete();
+        }
 
         $request->session()->invalidate();
         $request->session()->regenerateToken();
